+++ tests/testthat/test-plot.R
@@ -5,7 +5,13 @@ test_that("output is a ggplot", {
   
   secc <- load_geouy("Secciones")
   aaa <- plot_geouy(x = secc, col = "AREA")  
-  expect_is(aaa, "ggplot")
+  testthat::expect_is(aaa, "ggplot")
+  aaa <- plot_geouy(x = secc, col = "AREA", labels = "%")  
+  testthat::expect_is(aaa, "ggplot")
+  aaa <- plot_geouy(x = secc, col = "AREA", labels = "n")  
+  testthat::expect_is(aaa, "ggplot")
+  aaa <- plot_geouy(x = secc, col = "AREA", labels = "c", other_lab = "CODSEC")  
+  testthat::expect_is(aaa, "ggplot")
 })
 
 test_that("Plot uses correct data", {