+++ NEWS.md
@@ -3,7 +3,8 @@
 -------------------------------------------------------
 ## geouy v0.2.2 (2020-07-31)
 
-* Add loc_agr_ine dataset of "Localidades agragadas" of INE. 
+* Add loc_agr_ine dataset of "Localidades agragadas" of INE.
+* Add a geouy.R to discribe the package and set globals variables
 * replace ggsn with ggspatial to improve north and scale aesthetics in plot_geouy 
 * remove lat and long axis in plot_goeuy
 * Add a param labels in plot_geouy for labels posiblity, if "%" porcentage with 1 decimal labels, if "n" the value is the label, if "c" put other variable in other_lab